move get_generator to compiler.utils
